+++ src/completions.rs
@@ -13,34 +13,50 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 use lru::LruCache;
-use std::num::NonZeroUsize;
 
-const MAX_ENTRIES: usize = 32;
+const MAX_ENTRIES: usize = 40;
 
 /// A completions LRU cache.
 pub struct Completions {
-    completions: LruCache<String, ()>,
+    lru: LruCache<String, ()>,
 }
 
 impl Default for Completions {
     fn default() -> Self {
         Self {
-            completions: LruCache::new(NonZeroUsize::new(MAX_ENTRIES).unwrap()),
+            lru: LruCache::unbounded(),
         }
     }
 }
 
 impl Completions {
-    /// Add a new completion entry
-    pub fn add(&mut self, entry: &str) {
-        if self.completions.get(entry).is_none() {
-            self.completions.put(entry.to_string(), ());
+    /// Add entries to completions history.
+    pub fn add(&mut self, entries: &[String]) {
+        // Make sure these entries are in the cache irrespective of their size
+        // to handle the case where we have a dataframe with many columns.
+        if entries.len() > MAX_ENTRIES {
+            self.lru.clear();
+        } else if entries.len() + self.lru.len() > MAX_ENTRIES {
+            let to_remove = entries.len() + self.lru.len() - MAX_ENTRIES;
+            for _ in 0..to_remove {
+                self.lru.pop_lru();
+            }
+        }
+
+        for entry in entries {
+            self.add_entry(entry);
         }
     }
 
     /// Returns an iterator to the completions.
     pub fn iter(&self) -> impl Iterator<Item = &str> {
-        self.completions.iter().map(|(k, _)| k.as_str())
+        self.lru.iter().map(|(k, _)| k.as_str())
+    }
+
+    fn add_entry(&mut self, entry: &str) {
+        if self.lru.get(entry).is_none() {
+            self.lru.put(entry.to_string(), ());
+        }
     }
 }
 
@@ -52,21 +68,42 @@ mod tests {
     fn completions() {
         let mut completions = Completions::default();
 
-        for idx in 0..MAX_ENTRIES + 10 {
-            completions.add(&format!("entry{idx}"));
+        // This insertion should have all columns, simulates a dataframe that has
+        // more columns than MAX_ENTRIES.
+        let entries = (0..MAX_ENTRIES + 10)
+            .map(|idx| format!("entry{idx}"))
+            .collect::<Vec<_>>();
+
+        completions.add(&entries);
+        assert_eq!(completions.iter().count(), entries.len());
+
+        // Inserts a subsets of columns smaller than MAX_ENTRIES.
+        let entries = (1000..1020)
+            .map(|idx| format!("entry{idx}"))
+            .collect::<Vec<_>>();
+        completions.add(&entries);
+
+        assert_eq!(completions.iter().count(), MAX_ENTRIES);
+
+        // Most recently used are first.
+        for (entry, (cached, _)) in entries.iter().rev().zip(completions.lru.iter()) {
+            assert_eq!(entry, cached);
         }
 
-        // Reinsert same late entries
-        for idx in 20..MAX_ENTRIES + 10 {
-            completions.add(&format!("entry{idx}"));
-        }
+        let entries = (2000..2200)
+            .map(|idx| format!("entry{idx}"))
+            .collect::<Vec<_>>();
+        completions.add(&entries);
+        assert_eq!(completions.iter().count(), entries.len());
 
-        let items = completions.iter().collect::<Vec<_>>();
-        assert!(items.len() == MAX_ENTRIES);
+        let entries = (3000..3100)
+            .map(|idx| format!("entry{idx}"))
+            .collect::<Vec<_>>();
+        completions.add(&entries);
+        assert_eq!(completions.iter().count(), entries.len());
 
-        // First 10 should be evicted
-        for idx in 0..10 {
-            assert!(!items.contains(&format!("entry{idx}").as_str()));
+        for (entry, (cached, _)) in entries.iter().rev().zip(completions.lru.iter()) {
+            assert_eq!(entry, cached);
         }
     }
 }