+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Changes to the `dply` crate are documented in this file.
 ## 0.1.9 - Unreleased
 ### 🐛 Fixed
 * repl: Fix completions for absolute and tilde paths.
+* repl: Keep all completion columns for big dataframes.
 
 ## 0.1.8 - 2023-06-19
 ### ⭐ Added